Refactor simulation code into modules
