+++ threebody/constants.py
@@ -1,0 +1,74 @@
+# Constants for the N-body simulation
+import numpy as np
+
+# Application
+VERSION = "2.0.17"
+WIDTH, HEIGHT = 1280, 800
+UI_SIDEBAR_WIDTH = 300
+UI_BOTTOM_HEIGHT = 30
+
+# Colors
+BLACK = (0, 0, 0)
+DARK_GRAY = (20, 20, 20)
+GRAY = (80, 80, 80)
+LIGHT_GRAY = (160, 160, 160)
+WHITE = (255, 255, 255)
+
+SUN_COLOR = (255, 230, 110)
+MERCURY_COLOR = (200, 190, 190)
+VENUS_COLOR = (255, 190, 150)
+EARTH_COLOR = (100, 140, 255)
+MARS_COLOR = (220, 100, 60)
+JUPITER_COLOR = (255, 165, 60)
+GAS_COLOR = (230, 230, 210)
+ICE_COLOR = (160, 210, 255)
+STAR_COLORS = [
+    (255, 230, 180),
+    (255, 210, 130),
+    (255, 175, 100),
+    (255, 120, 80),
+]
+
+UI_BG_COLOR = (30, 30, 40)
+UI_ACCENT_COLOR = (60, 100, 180)
+UI_TEXT_COLOR = (230, 230, 230)
+
+# Physical constants
+G_REAL = 6.67430e-11
+AU = 1.496e11
+SOLAR_MASS = 1.989e30
+EARTH_MASS = 5.972e24
+JUPITER_MASS = 1.898e27
+EARTH_RADIUS_METERS = 6.371e6
+
+# Simulation parameters
+SPACE_SCALE = 5e9
+INITIAL_G = G_REAL
+TIME_STEP_BASE = 600
+SPEED_FACTOR = 1.0
+SOFTENING_FACTOR_SQ = 1.0 ** 2
+VELOCITY_DRAG_SCALE = 10000.0
+
+ADAPTIVE_STEPPING = True
+ERROR_TOLERANCE = 1e-6
+MIN_TIME_STEP = 10
+MAX_TIME_STEP = 3600
+
+FIELD_RESOLUTION = 30
+SHOW_GRAV_FIELD = False
+SHOW_TRAILS = True
+ZOOM_BASE = 1.0 / (AU / 500)
+INITIAL_PAN_OFFSET = np.array([(WIDTH - UI_SIDEBAR_WIDTH)/2,
+                               (HEIGHT - UI_BOTTOM_HEIGHT)/2],
+                              dtype=np.float64)
+CAMERA_SMOOTHING = 0.05
+ZOOM_FACTOR = 1.2
+BODY_ZOOM_SCALING_POWER = 0.2
+
+DEFAULT_NEXT_BODY_MASS = EARTH_MASS
+DEFAULT_NEXT_BODY_RADIUS_PIXELS = 8
+DEFAULT_TRAIL_LENGTH = 500
+MAX_DISPLAY_BODIES = 100
+COLLISION_DISTANCE_FACTOR = 5.0
+
+DEBUG_PHYSICS_COUNT = 0
